test_shell에 testapp unittest 추가 (기존에 임의로 testapp unittest 개발한 파일 삭제)
